Move fuzzy_date and modal_workflow from lib to common.utils
