+++ comments/tests/utils.py
@@ -2,8 +2,8 @@ import datetime
 
 from django.contrib.auth.models import User
 from django.test import TestCase
-from fuzzy_date import FuzzyDate
 
+from common.utils.fuzzy_date import FuzzyDate
 from comments.models import Comment
 from parties.models import Party, PartySeries
 from productions.models import Production
