+++ demoscene/tests/test_tags.py
@@ -1,7 +1,7 @@
 from django import forms, template
 from django.test import TestCase
-from fuzzy_date import FuzzyDate
 
+from common.utils.fuzzy_date import FuzzyDate
 from demoscene.templatetags.demoscene_tags import date_range
 from demoscene.templatetags.safe_markdown import safe_markdown
