+++ lib/fuzzy_date_field.py
@@ -2,7 +2,7 @@ from django import forms
 from django.core import validators
 from django.core.exceptions import ValidationError
 from django.utils.translation import gettext_lazy as _
-from fuzzy_date import FuzzyDate
+from common.utils.fuzzy_date import FuzzyDate
 
 
 class FuzzyDateField(forms.Field):
