+++ parties/models.py
@@ -8,7 +8,6 @@ from django.core.files.storage import FileSystemStorage
 from django.db import models
 from django.db.models import Prefetch
 from django.urls import reverse
-from fuzzy_date import FuzzyDate
 from strip_markup import strip_markup
 from unidecode import unidecode
 
@@ -16,6 +15,7 @@ from comments.models import Commentable
 from demoscene.models import DATE_PRECISION_CHOICES, ExternalLink, Releaser, TextFile
 from common.utils import groklinks
 from common.utils.files import random_path
+from common.utils.fuzzy_date import FuzzyDate
 from common.utils.text import generate_search_title
 from productions.models import Production, Screenshot
