+++ parties/tests/test_models.py
@@ -2,8 +2,8 @@ import datetime
 
 from django.test import TestCase
 from freezegun import freeze_time
-from fuzzy_date import FuzzyDate
 
+from common.utils.fuzzy_date import FuzzyDate
 from parties.models import Party, PartySeries
 from productions.models import Production
 from sceneorg.models import Directory, File
