+++ productions/models.py
@@ -11,7 +11,6 @@ from django.dispatch import receiver
 from django.forms import Media
 from django.urls import reverse
 from django.utils.translation import gettext_lazy as _
-from fuzzy_date import FuzzyDate
 from taggit.managers import TaggableManager
 from treebeard.mp_tree import MP_Node
 from unidecode import unidecode
@@ -19,6 +18,7 @@ from unidecode import unidecode
 from comments.models import Commentable
 from demoscene.models import DATE_PRECISION_CHOICES, ExternalLink, Nick, Releaser, ReleaserExternalLink, TextFile
 from common.utils import groklinks
+from common.utils.fuzzy_date import FuzzyDate
 from common.utils.text import generate_search_title, generate_sort_key
 from lib.lockable import Lockable
 from lib.prefetch_snooping import ModelWithPrefetchSnooping
